requirements.txt
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-mecab-ko>=1.3.0
matplotlib>=3.6.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
requests>=2.28.0
# Microsoft Teams/Graph API 패키지
msal>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.0 